chore: move table hooks to dedicated hooks folder
