+++ web/src/features/alerts/pages/alert-rules-page/AlertRulesPage.tsx
@@ -46,9 +46,12 @@ import {
 } from '../../../../shared/components/table/TableFilters';
 import { TableToolbar } from '../../../../shared/components/table/TableToolbar';
 import { directedSort, localeCompareSort } from '../../../../shared/components/table/sort-utils';
-import { useTableColumns } from '../../../../shared/components/table/useTableColumns';
-import { useTableFilters, rowFilter } from '../../../../shared/components/table/useTableFilters';
-import { useTablePagination } from '../../../../shared/components/table/useTablePagination';
+import { useTableColumns } from '../../../../shared/components/table/hooks/useTableColumns';
+import {
+  useTableFilters,
+  rowFilter,
+} from '../../../../shared/components/table/hooks/useTableFilters';
+import { useTablePagination } from '../../../../shared/components/table/hooks/useTablePagination';
 import { filterRules } from './filter-rules';
 import { SeverityBadge } from '../../../../shared/components/SeverityBadge';
