+++ web/src/features/alerts/pages/silences-page/SilencesPage.tsx
@@ -49,7 +49,7 @@ import {
   usePerspective,
 } from '../../../../shared/hooks/usePerspective';
 import { directedSort, localeCompareSort } from '../../../../shared/components/table/sort-utils';
-import { useTableColumns } from '../../../../shared/components/table/useTableColumns';
+import { useTableColumns } from '../../../../shared/components/table/hooks/useTableColumns';
 import {
   ITEMS_PER_PAGE,
   TablePagination,
@@ -61,8 +61,11 @@ import {
   TableFilters,
 } from '../../../../shared/components/table/TableFilters';
 import { TableToolbar } from '../../../../shared/components/table/TableToolbar';
-import { useTableFilters, rowFilter } from '../../../../shared/components/table/useTableFilters';
-import { useTablePagination } from '../../../../shared/components/table/useTablePagination';
+import {
+  useTableFilters,
+  rowFilter,
+} from '../../../../shared/components/table/hooks/useTableFilters';
+import { useTablePagination } from '../../../../shared/components/table/hooks/useTablePagination';
 import { severitySort, SilenceResource, silenceState } from '../../../../shared/utils/utils';
 import { SeverityCounts, StateTimestamp } from '../../components/AlertUtils';
 import { filterSilences } from './filter-silences';
