+++ web/src/features/targets/pages/targets-page.tsx
@@ -56,7 +56,7 @@ import { usePoll } from '../../../shared/console/utils/poll-hook';
 import { useSafeFetch } from '../../../shared/console/utils/safe-fetch-hook';
 import { filterTargets } from '../utils/filter-targets';
 import { directedSort, localeCompareSort } from '../../../shared/components/table/sort-utils';
-import { useTableColumns } from '../../../shared/components/table/useTableColumns';
+import { useTableColumns } from '../../../shared/components/table/hooks/useTableColumns';
 import { useBoolean } from '../../../shared/hooks/useBoolean';
 import { Labels } from '../../../shared/components/labels';
 import { ITEMS_PER_PAGE, TablePagination } from '../../../shared/components/table/table-pagination';
@@ -67,8 +67,8 @@ import {
   TableFilters,
 } from '../../../shared/components/table/TableFilters';
 import { TableToolbar } from '../../../shared/components/table/TableToolbar';
-import { useTableFilters, rowFilter } from '../../../shared/components/table/useTableFilters';
-import { useTablePagination } from '../../../shared/components/table/useTablePagination';
+import { useTableFilters, rowFilter } from '../../../shared/components/table/hooks/useTableFilters';
+import { useTablePagination } from '../../../shared/components/table/hooks/useTablePagination';
 import { AlertSource, PrometheusAPIError, Target } from '../../../shared/types/types';
 import { PROMETHEUS_BASE_PATH } from '../../../shared/utils/utils';
