+++ web/src/shared/components/table/hooks/useTableFilters.ts
